requirements.txt
Django==3.2.12
djangorestframework==3.13.1
django-bootstrap5==0.5.0
django-redis==5.1.0
requests==2.26.0


